Moving object model classes
